@@ -18,7 +18,7 @@ classifiers =
 package_dir =
     = .
 packages = find:
-python_requires = >=3.8
+python_requires = >=3.10
 install_requires =
     pyyaml
     cython<3.0
