@@ -1,7 +1,7 @@
 from pyodide_test_runner import run_in_pyodide
 
 
-@run_in_pyodide(standalone=True, packages=["msgpack"])
+@run_in_pyodide(selenium_fixture_name="selenium_standalone", packages=["msgpack"])
 def test_pack():
     from msgpack import Packer, Unpacker, pack, packb, unpackb  # noqa: F401
 
