@@ -3,7 +3,9 @@ from pyodide_test_runner import run_in_pyodide as run_in_pyodide_orig
 
 
 def run_in_pyodide(**kwargs):
-    return run_in_pyodide_orig(module_scope=True, packages=["numpy"])
+    return run_in_pyodide_orig(
+        selenium_fixture_name="selenium_module_scope", packages=["numpy"]
+    )
 
 
 def test_numpy(selenium):
@@ -194,11 +196,11 @@ def test_runpythonasync_numpy(selenium_standalone):
         )
 
 
+@pytest.mark.xfail_browsers(
+    firefox="Timeout in WebWorker when using numpy in Firefox 87"
+)
 @pytest.mark.driver_timeout(30)
 def test_runwebworker_numpy(selenium_webworker_standalone):
-    if selenium_webworker_standalone.browser == "firefox":
-        pytest.xfail("Timeout in WebWorker when using numpy in Firefox 87")
-
     output = selenium_webworker_standalone.run_webworker(
         """
         import numpy as np
