@@ -1,9 +1,10 @@
 import random
+from typing import Any
 
 import pytest
 
 
-def generate_largish_json(n_rows: int = 91746) -> dict:
+def generate_largish_json(n_rows: int = 91746) -> dict[str, Any]:
     # with n_rows = 91746, the output JSON size will be ~15 MB/10k rows
 
     # Note: we don't fix the random seed here, but the actual values
@@ -40,17 +41,14 @@ def test_extra_import(selenium, request):
     selenium.run("from pandas import Series, DataFrame")
 
 
+@pytest.mark.xfail_browsers(
+    chrome="test_load_largish_file triggers a fatal runtime error in Chrome 89 see #1495",
+    node="open_url doesn't work in node",
+)
 @pytest.mark.driver_timeout(40)
 @pytest.mark.skip_refcount_check
 def test_load_largish_file(selenium_standalone, request, httpserver):
     selenium = selenium_standalone
-    if selenium.browser == "chrome":
-        pytest.xfail(
-            "test_load_largish_file triggers a fatal runtime error in Chrome 89 see #1495"
-        )
-    if selenium.browser == "node":
-        pytest.xfail("open_url doesn't work in node")
-
     selenium.load_package("pandas")
     selenium.load_package("matplotlib")
 
