@@ -26,13 +26,20 @@ def set_webdriver_script_timeout(selenium, script_timeout: float | None):
         selenium.set_script_timeout(selenium.script_timeout)
 
 
-def parse_driver_timeout(request) -> float | None:
+def parse_driver_timeout(node) -> float | None:
     """Parse driver timeout value from pytest request object"""
-    mark = request.node.get_closest_marker("driver_timeout")
+    mark = node.get_closest_marker("driver_timeout")
     if mark is None:
         return None
     else:
         return mark.args[0]
+
+
+def parse_xfail_browsers(node) -> dict[str, str]:
+    mark = node.get_closest_marker("xfail_browsers")
+    if mark is None:
+        return {}
+    return mark.kwargs
 
 
 def maybe_skip_test(item, dist_dir, delayed=False):
