@@ -14,11 +14,10 @@ def get_pytz_wheel_name():
     return list(DIST_PATH.glob("pytz*.whl"))[0].name
 
 
+@pytest.mark.xfail_browsers(node="Loading urls in node seems to time out right now")
 @pytest.mark.parametrize("active_server", ["main", "secondary"])
 def test_load_from_url(selenium_standalone, web_server_secondary, active_server):
     selenium = selenium_standalone
-    if selenium.browser == "node":
-        pytest.xfail("Loading urls in node seems to time out right now")
     if active_server == "secondary":
         url, port, log_main = web_server_secondary
         log_backup = selenium.server_log
@@ -313,7 +312,7 @@ def test_load_bad_so_file(selenium):
     selenium.run_js(
         """
         pyodide.FS.writeFile("/a.so", new Uint8Array(4))
-        await pyodide._api.tests.loadDynlib("/a.so");
+        await pyodide._api.loadDynlib("/a.so");
         """
     )
     assert (
@@ -385,11 +384,11 @@ def test_get_dynlibs():
             x.addfile(tarfile.TarInfo(file))
         x.close()
         t.flush()
-        assert sorted(get_dynlibs(t, Path("/p"))) == so_files
+        assert sorted(get_dynlibs(t, ".bz", Path("/p"))) == so_files
     with NamedTemporaryFile(suffix=".zip") as t:
         x2 = ZipFile(t, mode="w")
         for file in files:
             x2.writestr(file, "")
         x2.close()
         t.flush()
-        assert sorted(get_dynlibs(t, Path("/p"))) == so_files
+        assert sorted(get_dynlibs(t, ".zip", Path("/p"))) == so_files
